Hanukkah
I had to do some corrections involving spacing and the like. I also added information (both in comments and in dialog) on why the days shift from year to year.

(Very low priority commentary here)
There is also commented information for semi-far future notice as calendar holidays appear to now be in process of being implemented (probably in phases). There is a possibility that appropriate calendar conversions are being placed. - I put a specific phrase called "Hebrew Lunar Calendar." It may be recommended that if there are future holidays that shift around the Gregorian Calendar due to the use of a different one (Easter, Passover, Ramadan, etc.) to have such commented references that it uses a specific calendar (the complete name/type of calendar) and the name of the month that it takes place. (e.g. Ramadan is a month of the Islamic Calendar.)